Home

StorageGraphs

StorageGraphs is a package for storing hierarchical data in graphs in a non-redundant way. This package aims to be useful when one has a combination of data and metadata or parameters and the use of tables would lead to a lot of redundancy in the corresponding columns.

StorageGraphs

StorageGraphs.add_bulk! — Method.

add_bulk!(g, dep, vals)

Add the multiple values (vals) of the things identified by the keys of vals, with the dependency chain given by dep. The values of vals are assumed to be vectors. Each added node will correspond to an element of the vectors. Note: The dependency chain must contain all relevant information for identifying the values.

source

StorageGraphs.add_derived_values! — Method.

add_derived_values!(g, base_dep, base_val, val, inner_deps...)

Add multiple values such that the elements in base_val and val are linked in such a way that the order is preserved. This is useful when one wants to add a vector of values derived from another vector. The dependency for the base values (base_dep) must be given as a collection of NamedTuples instead of a nested Pair. Also, any inner dependencies (inner_deps) must be given as individual NamedTuples. A new path is created for each value, but if a part already exists, it is continued (see nextid).

source

StorageGraphs.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=nextid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one, but if there is an existing part, it is continued (see nextid).

source

StorageGraphs.add_path! — Method.

add_path!(g, source, dest; id=maxid(g))

Create a path between the source node and the destination one. If the nodes do not exist, they are created.

source

StorageGraphs.final_neighborhs — Method.

final_neighborhs(g, dep::Pair; dir=:out)

Return the vertex indices for the neighbors at the end of the dependency chain. Note: this assumes that the dependency chain is valid (all the nodes exist).

source

StorageGraphs.findnodes — Method.

findnodes(g, name::Symbol)

Finds the nodes containing name.

source

StorageGraphs.paths_through — Method.

paths_through(g, v::Integer; dir=:out)

Return a vector of the paths going through the given vertex. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

StorageGraphs.nextid — Method.

nextid(g, dep::Pair)

Find the next available id such that a dead end (a node with no outgoing paths) along the dependency chain (dep) is continued. If there is no such case, it gives the maximum id (see walkdep).

source

StorageGraphs.on_path — Method.

on_path(g, v, path)

Check if the vertex is on the given path.

source

StorageGraphs.walkdep — Method.

function walkdep(g, dep::Pair; stopcond=(g,v)->false)

Walk along the dependency chain, but only on already existing paths, and return the last node and the compatible paths.

source

StorageGraphs.walkpath — Method.

walkpath(g, paths, start; dir=:out, stopcond=(g,v)->false)

Walk on the given paths starting from start and return the last nodes. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source