Adding data

Adding data

StorageGraphs.add_bulk! — Method.

add_bulk!(g, dep, vals)

Add the multiple values (vals) of the things identified by the keys of vals, with the dependency chain given by dep. The values of vals are assumed to be equal length vectors. Each added node will correspond to an element of the vectors. Note: The dependency chain must contain all relevant information for identifying the values.

The function returns the ids of the paths corresponding to the added values.

source

StorageGraphs.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=nextid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one, but if there is an existing part, it is continued (see nextid).

source