Internals

Index

LightGraphs.SimpleGraphs.add_edge!

LightGraphs.SimpleGraphs.add_vertex!

StorageGraphs.add_bulk!

StorageGraphs.add_derived_values!

StorageGraphs.add_nodes!

StorageGraphs.add_path!

StorageGraphs.extractvals

StorageGraphs.final_neighborhs

StorageGraphs.findnodes

StorageGraphs.get_prop

StorageGraphs.has_prop

StorageGraphs.nextid

StorageGraphs.on_path

StorageGraphs.ordered_dependency

StorageGraphs.paths_through

StorageGraphs.rem_prop!

StorageGraphs.set_prop!

StorageGraphs.walkdep

StorageGraphs.walkpath

StorageGraphs.walkpath!

Internal API

LightGraphs.SimpleGraphs.add_edge! — Method.

add_edge!(g, u, v, id)

Add an edge (u, v) to the StorageGraph g with the given id. Return true if the edge has been added, false otherwise.

source

LightGraphs.SimpleGraphs.add_vertex! — Method.

add_vertex!(g)
add_vertex!(g, data)

Add a vertex to the StorageGraph g with optional data given by the NamedTuple data. Return true if the vertex has been added, false otherwise.

source

StorageGraphs.add_path! — Method.

add_path!(g, source, dest; id=maxid(g))

Create a path between the source node and the destination one. If the nodes do not exist, they are created.

source

StorageGraphs.extractvals — Method.

extractvals(nodes, name::Symbol)

Return an array of values corresponding to name form the array of NamedTuples nodes.

source

StorageGraphs.findnodes — Method.

findnodes(g, name::Symbol)

Finds the nodes containing name.

source

StorageGraphs.ordered_dependency — Method.

ordered_dependency(a, b, inner_deps...)

Return a vector of dependency chains such that the elements of a are linked to the ones in b in such a way that the order is preserved.

source

StorageGraphs.rem_prop! — Method.

rem_prop!(g, v)
rem_prop!(g, e)
rem_prop!(g, s, d)

Remove the specific property (data for vertices, ids for edges) from graph g, vertex v, or edge e (optionally referenced by source vertex s and destination vertex d). If property, vertex, or edge does not exist, will not do anything.

source

StorageGraphs.walkpath! — Method.

walkpath!(g, path, start, neighborfn, action!; stopcond=(g,v)->false)

Walk on the given path and take an action at each node. The action is specified by a function action!(g, v, neighbors) and it can modify the graph.

source