Home

GraphStorage.jl

GraphStorage.add_bulk!

GraphStorage.add_derived_values!

GraphStorage.add_node!

GraphStorage.add_nodes!

GraphStorage.add_path!

GraphStorage.extractvals

GraphStorage.final_neighborhs

GraphStorage.findnodes

GraphStorage.nextid

GraphStorage.nodevals

GraphStorage.on_path

GraphStorage.ordered_dependency

GraphStorage.paths_through

GraphStorage.walkdep

GraphStorage.walkpath

GraphStorage.walkpath!

GraphStorage.add_bulk! — Method.

add_bulk!(g, dep, vals)

Add the multiple values (vals) of the things identified by the keys of vals, with the dependency chain given by dep. The values of vals are assumed to be vectors. Each added node will correspond to an element of the vectors. Note: The dependency chain must contain all relevant information for identifying the values.

source

GraphStorage.add_derived_values! — Method.

add_derived_values!(g, base_dep, base_val, val, inner_deps...)

Add multiple values such that the elements in base_val and val are linked in such a way that the order is preserved. This is useful when one wants to add a vector of values derived from another vector. The dependency for the base values (base_dep) must be given as a collection of NamedTuples instead of a nested Pair. Also, any inner dependencies (inner_deps) must be given as individual NamedTuples. A new path is created for each value, but if a part already exists, it is continued (see nextid).

source

GraphStorage.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=nextid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one, but if there is an existing part, it is continued (see nextid).

source

GraphStorage.add_path! — Method.

add_path!(g, source, dest; id=maxid(g))

Create a path between the source node and the destination one. If the nodes do not exist, they are created.

source

GraphStorage.final_neighborhs — Method.

final_neighborhs(g, dep::Pair; dir=:out)

Return the vertex indices for the neighbors at the end of the dependency chain.

source

GraphStorage.findnodes — Method.

findnodes(g, name::Symbol)

Finds the nodes containing name.

source

GraphStorage.nextid — Method.

nextid(g, dep::Pair)

Find the next available id such that a dead end (a node with no outgoing paths) along the dependency chain (dep) is continued. If there is no such case, it gives the maximum id (see walkdep).

source

GraphStorage.nodevals — Method.

nodevals(g, name::Symbol)

Return an array of the values corresponding to name. See also findnodes.

source

GraphStorage.on_path — Method.

on_path(g, v, path)

Check if the vertex is on the given path.

source

GraphStorage.paths_through — Method.

paths_through(g, v::Integer; dir=:out)

Return a vector of the paths going through the given vertex. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

GraphStorage.walkdep — Method.

function walkdep(g, dep::Pair; stopcond=(g,v)->false)

Walk along the dependency chain, but only on already existing paths, and return the last node and the compatible paths.

source

GraphStorage.walkpath — Method.

walkpath(g, paths, start; dir=:out, stopcond=(g,v)->false)

Walk on the given paths starting from start and return the last nodes. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

GraphStorage.add_node! — Method.

add_node!(g, val::NamedTuple)

Add a new node to the storage graph.

source

GraphStorage.extractvals — Method.

extractvals(nodes, name)

Return an array of values corresponding to name form the array of NamedTuples nodes.

source

GraphStorage.ordered_dependency — Method.

ordered_dependency(a, b, inner_deps...)

Return a vector of dependency chains such that the elements of a are linked to the ones in b in such a way that the order is preserved.

source

GraphStorage.walkpath! — Method.

walkpath!(g, path, start, neighborfn, action!; stopcond=(g,v)->false)

Walk on the given path and take an action at each node. The action is specified by a function action!(g, v, neighbors) and it can modify the graph.

source