Home

GraphStorage.jl

GraphStorage.add_node!

GraphStorage.add_nodes!

GraphStorage.add_path!

GraphStorage.get_node_index

GraphStorage.indexby

GraphStorage.on_path

GraphStorage.paths_through

GraphStorage.walkpath

GraphStorage.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=maxid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one.

source

GraphStorage.add_path! — Method.

add_path!(g, source, dest; id=maxid(g))

Create a path between the source node and the destination one. If the nodes do not exist, they are created.

source

GraphStorage.get_node_index — Method.

get_node_index(g, val; createnew=true)

Get the index of a node identified by a NamedTuple. If it doesn't exist, it can be created.

source

GraphStorage.indexby — Method.

indexby(g, key)

Set key as an indexing property.

source

GraphStorage.on_path — Method.

on_path(g, v, path)

Check if the vertex is on the given path.

source

GraphStorage.paths_through — Method.

paths_through(g, v::Integer; dir=:out)

Return a vector of the paths going through the given vertex. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

GraphStorage.walkpath — Method.

walkpath(g, paths, start; dir=:out, stopcond=(g,v)->true)

Walk on the given paths starting from start and return the last nodes. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

GraphStorage.add_node! — Method.

add_node!(g, val::NamedTuple)

Add a new node to the storage graph.

source