Home

StorageGraphs.jl

LightGraphs.SimpleGraphs.add_edge!

LightGraphs.SimpleGraphs.add_vertex!

StorageGraphs.add_bulk!

StorageGraphs.add_derived_values!

StorageGraphs.add_nodes!

StorageGraphs.add_path!

StorageGraphs.extractvals

StorageGraphs.final_neighborhs

StorageGraphs.findnodes

StorageGraphs.get_prop

StorageGraphs.nextid

StorageGraphs.nodevals

StorageGraphs.on_path

StorageGraphs.ordered_dependency

StorageGraphs.paths_through

StorageGraphs.rem_prop!

StorageGraphs.set_prop!

StorageGraphs.walkdep

StorageGraphs.walkpath

StorageGraphs.walkpath!

StorageGraphs.add_bulk! — Method.

add_bulk!(g, dep, vals)

Add the multiple values (vals) of the things identified by the keys of vals, with the dependency chain given by dep. The values of vals are assumed to be vectors. Each added node will correspond to an element of the vectors. Note: The dependency chain must contain all relevant information for identifying the values.

source

StorageGraphs.add_derived_values! — Method.

add_derived_values!(g, base_dep, base_val, val, inner_deps...)

Add multiple values such that the elements in base_val and val are linked in such a way that the order is preserved. This is useful when one wants to add a vector of values derived from another vector. The dependency for the base values (base_dep) must be given as a collection of NamedTuples instead of a nested Pair. Also, any inner dependencies (inner_deps) must be given as individual NamedTuples. A new path is created for each value, but if a part already exists, it is continued (see nextid).

source

StorageGraphs.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=nextid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one, but if there is an existing part, it is continued (see nextid).

source

StorageGraphs.add_path! — Method.

add_path!(g, source, dest; id=maxid(g))

Create a path between the source node and the destination one. If the nodes do not exist, they are created.

source

StorageGraphs.final_neighborhs — Method.

final_neighborhs(g, dep::Pair; dir=:out)

Return the vertex indices for the neighbors at the end of the dependency chain.

source

StorageGraphs.findnodes — Method.

findnodes(g, name::Symbol)

Finds the nodes containing name.

source

StorageGraphs.get_prop — Method.

get_prop(g, v)
get_prop(g, e)
get_prop(g, s, d)

Return the specific property (data for vertices, ids for edges) defined for graph g, vertex v, or edge e (optionally referenced by source vertex s and destination vertex d). If property is not defined, return an error.

source

StorageGraphs.nextid — Method.

nextid(g, dep::Pair)

Find the next available id such that a dead end (a node with no outgoing paths) along the dependency chain (dep) is continued. If there is no such case, it gives the maximum id (see walkdep).

source

StorageGraphs.nodevals — Method.

nodevals(g, name::Symbol)

Return an array of the values corresponding to name. See also findnodes.

source

StorageGraphs.on_path — Method.

on_path(g, v, path)

Check if the vertex is on the given path.

source

StorageGraphs.paths_through — Method.

paths_through(g, v::Integer; dir=:out)

Return a vector of the paths going through the given vertex. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

StorageGraphs.set_prop! — Method.

set_prop!(g, val)
set_prop!(g, v, val)
set_prop!(g, e, val)
set_prop!(g, s, d, val)

Set (replace) the specific property (data for vertices, ids for edges, max id for the graph) with value val in graph g, vertex v, or edge e (optionally referenced by source vertex s and destination vertex d). Will return false if vertex or edge does not exist, true otherwise.

source

StorageGraphs.walkdep — Method.

function walkdep(g, dep::Pair; stopcond=(g,v)->false)

Walk along the dependency chain, but only on already existing paths, and return the last node and the compatible paths.

source

StorageGraphs.walkpath — Method.

walkpath(g, paths, start; dir=:out, stopcond=(g,v)->false)

Walk on the given paths starting from start and return the last nodes. If dir is specified, use the corresponding edge direction (:in and :out are acceptable values).

source

LightGraphs.SimpleGraphs.add_edge! — Method.

add_edge!(g, u, v, d)

Add an edge (u, v) to the StorageGraph g with the given id. Return true if the edge has been added, false otherwise.

source

LightGraphs.SimpleGraphs.add_vertex! — Method.

add_vertex!(g)
add_vertex!(g, data)

Add a vertex to the StorageGraph g with optional data given by the NamedTuple data. Return true if the vertex has been added, false otherwise.

source

StorageGraphs.extractvals — Method.

extractvals(nodes, name)

Return an array of values corresponding to name form the array of NamedTuples nodes.

source

StorageGraphs.ordered_dependency — Method.

ordered_dependency(a, b, inner_deps...)

Return a vector of dependency chains such that the elements of a are linked to the ones in b in such a way that the order is preserved.

source

StorageGraphs.rem_prop! — Method.

rem_prop!(g, v)
rem_prop!(g, e)
rem_prop!(g, s, d)

Remove the specific property (data for vertices, ids for edges) from graph g, vertex v, or edge e (optionally referenced by source vertex s and destination vertex d). If property, vertex, or edge does not exist, will not do anything.

source

StorageGraphs.walkpath! — Method.

walkpath!(g, path, start, neighborfn, action!; stopcond=(g,v)->false)

Walk on the given path and take an action at each node. The action is specified by a function action!(g, v, neighbors) and it can modify the graph.

source