Adding data

Adding data

StorageGraphs.add_bulk! — Method.

add_bulk!(g, dep, vals)

Add the multiple values (vals) of the things identified by the keys of vals, with the dependency chain given by dep. The values of vals are assumed to be vectors. Each added node will correspond to an element of the vectors. Note: The dependency chain must contain all relevant information for identifying the values.

source

StorageGraphs.add_derived_values! — Method.

add_derived_values!(g, base_dep, base_val, val, inner_deps...)

Add multiple values such that the elements in base_val and val are linked in such a way that the order is preserved. This is useful when one wants to add a vector of values derived from another vector. The dependency for the base values (base_dep) must be given as a collection of NamedTuples instead of a nested Pair. Also, any inner dependencies (inner_deps) must be given as individual NamedTuples. A new path is created for each value, but if a part already exists, it is continued (see nextid).

source

StorageGraphs.add_nodes! — Method.

add_nodes!(g, dep::Pair; id=nextid(g))

Recursively add nodes via the dependency chain specified by dep. If any intermediarry node doesn't exist, it is created. A new path is created starting from the first node to the last one, but if there is an existing part, it is continued (see nextid).

source